deps check

+++ certbot-manual-authenticator.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+./check-dependencies.sh
+if [[ "$?" -ne 0 ]]; then
+    exit 1
+fi
+
 if [ -f /tmp/certbot-dnsmasq.pid ]; then
 	kill $(cat /tmp/certbot-dnsmasq.pid)
 fi
